+++ CHANGES.rst
@@ -4,6 +4,7 @@ Version 1.0 (dev)
 * Added frontend media support.
 * **API Change:** Renamed template tag library ``placeholder_tags`` to ``fluent_contents_tags`` (the old name still works).
 * **API Change:** ``render_placeholder()`` and ``render_content_items()`` return a ``ContentItemOutput`` object, which can be treated like a string.
+* Content items are prefixed with "content:" during syncdb, a ``prefix_content_item_types`` management command can be run manually too.
 * Removed unneeded ``render_comment_list`` templatetag as it was upstreamed to django-threadedcomments_ 0.9.